+++ examples/obstacles.py
@@ -1,13 +1,13 @@
 """Demonstrate implemented features using renderer."""
 
-import logging
+from logging_config import configure_logger
 
 from two_d_game_ai.entities.bot import Bot
 from two_d_game_ai.entities.obstacles import ObstacleCircle, ObstacleRectangle
 from two_d_game_ai.view.view import View
 from two_d_game_ai.world.world import World
 
-logging.basicConfig(level=logging.INFO)
+configure_logger()
 
 # Create a World
 the_world = World(