+++ examples/zombies.py
@@ -3,14 +3,15 @@
 Zombies chase the human if they spot them, and stop chasing if they lose sight of them.
 """
 
-import logging
 import random
+
+from logging_config import configure_logger
 
 from two_d_game_ai.entities.bot import Bot
 from two_d_game_ai.view.view import View
 from two_d_game_ai.world.world import World
 
-logging.basicConfig(level=logging.DEBUG)
+configure_logger()
 
 ZOMBIE_MIN_SPEED = 1
 ZOMBIE_MAX_SPEED = 8